# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      success:
        type: boolean
    type: object
  models.Image:
    properties:
      content_type:
        type: string
      created_at:
        type: string
      etag:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      position:
        description: Position in user's photo gallery (1-9)
        type: integer
      s3_key:
        type: string
      size:
        type: integer
      thumbnail_url:
        type: string
      updated_at:
        type: string
      url:
        type: string
      user_id:
        type: string
    type: object
  models.ImageUploadBase64Request:
    properties:
      filename:
        type: string
      image_data:
        description: base64 encoded
        type: string
      position:
        maximum: 9
        minimum: 1
        type: integer
    required:
    - image_data
    type: object
  models.ImageUploadResponse:
    properties:
      content_type:
        type: string
      image_id:
        type: string
      position:
        type: integer
      s3_key:
        type: string
      size:
        type: integer
      thumbnail_url:
        type: string
      url:
        type: string
    type: object
...
  models.PresignedUploadRequest:
    properties:
      content_type:
        type: string
      position:
        maximum: 9
        minimum: 1
        type: integer
    required:
    - content_type
    type: object
  models.PresignedUploadResponse:
    properties:
      expires_at:
        type: string
      image_id:
        type: string
      public_url:
        type: string
      s3_key:
        type: string
      thumbnail_url:
        type: string
      upload_url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@matching-api.example.com
    name: API Support
    url: https://matching-api.example.com/support
  description: A complete Go-based REST API for a dating application with real-time
    chat, sophisticated matching algorithms, and comprehensive user management.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://matching-api.example.com/terms
  title: Matching API - Matching App Backend
  version: "1.0"
paths:
...
  /images:
    get:
      consumes:
      - application/json
      description: Get all images uploaded by the current user
      produces:
      - application/json
      responses:
        "200":
          description: Images retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  properties:
                    images:
                      items:
                        $ref: '#/definitions/models.Image'
                      type: array
                    total:
                      type: integer
                  type: object
              type: object
        "401":
          description: Unauthorized - invalid token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List user images
      tags:
      - Images
  /images/{imageKey}:
    delete:
      consumes:
      - application/json
      description: Delete an image owned by the user
      parameters:
      - description: S3 image key
        in: path
        name: imageKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Image deleted successfully
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad request - missing image key
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized - invalid token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Image not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete image
      tags:
      - Images
  /images/download/{imageKey}:
    get:
      consumes:
      - application/json
      description: Download an image file owned by the user
      parameters:
      - description: S3 image key
        in: path
        name: imageKey
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Image file
          schema:
            type: file
        "400":
          description: Bad request - missing image key
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized - invalid token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Image not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Download image
      tags:
      - Images
  /images/presigned-upload:
    post:
      consumes:
      - application/json
      description: Generate a presigned URL for direct image upload to S3
      parameters:
      - description: Presigned upload request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.PresignedUploadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Presigned URL generated successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.PresignedUploadResponse'
              type: object
        "400":
          description: Bad request - invalid content type
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized - invalid token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Generate presigned upload URL
      tags:
      - Images
  /images/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image file to user's profile
      parameters:
      - description: Image file (JPEG/PNG/WebP, max 10MB)
        in: formData
        name: image
        required: true
        type: file
      - default: 1
        description: Image position (1-9)
        in: query
        name: position
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Image uploaded successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ImageUploadResponse'
              type: object
        "400":
          description: Bad request - invalid file
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized - invalid token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload image
      tags:
      - Images
  /images/upload-base64:
    post:
      consumes:
      - application/json
      description: Upload an image from base64 encoded data
      parameters:
      - description: Base64 image data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ImageUploadBase64Request'
      produces:
      - application/json
      responses:
        "201":
          description: Base64 image uploaded successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ImageUploadResponse'
              type: object
        "400":
          description: Bad request - invalid base64 data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized - invalid token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload base64 image
      tags:
      - Images
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"